+++ core/src/main/java/nl/nn/adapterframework/util/XmlUtils.java
@@ -39,6 +39,7 @@ import java.util.Map;
 import java.util.Set;
 import java.util.StringTokenizer;
 
+import javax.xml.XMLConstants;
 import javax.xml.namespace.QName;
 import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
@@ -420,6 +421,7 @@ public class XmlUtils {
 
 		DocumentBuilderFactory factory = getDocumentBuilderFactory(xslt2, namespaceAware);
 		try {
+			factory.setFeature(XMLConstants.FEATURE_SECURE_PROCESSING, true);
 			DocumentBuilder builder = factory.newDocumentBuilder();
 			if (!resolveExternalEntities) {
 				builder.setEntityResolver(new XmlExternalEntityResolver());
@@ -765,6 +767,7 @@ public class XmlUtils {
 		InputSource is = in.asXmlInputSource();
 		SAXParserFactory factory = getSAXParserFactory(namespaceAware);
 		try {
+			factory.setFeature(XMLConstants.FEATURE_SECURE_PROCESSING, true);
 			XMLReader xmlReader = factory.newSAXParser().getXMLReader();
 			if (!resolveExternalEntities) {
 				xmlReader.setEntityResolver(new XmlExternalEntityResolver());
@@ -1452,6 +1455,18 @@ public class XmlUtils {
 		return result;
 	}
 
+	static public String identityTransform(String input)
+			throws DomBuilderException {
+		String result = "";
+		Document document = XmlUtils.buildDomDocument((String) input);
+		try {
+			result = nodeToString(document, false);
+		} catch (TransformerException e) {
+			throw new DomBuilderException(e);
+		}
+		return result;
+	}
+	
 	public static String getVersionInfo() {
 		StringBuilder sb = new StringBuilder();
 		sb.append(AppConstants.getInstance().getProperty("application.name") + " "
@@ -1617,8 +1632,14 @@ public class XmlUtils {
 	}
 
 	public static String nodeToString(Node node) throws TransformerException {
+		return nodeToString(node, true);
+	}
+
+	public static String nodeToString(Node node, boolean omitXmlDeclaration) throws TransformerException {
 		Transformer t = getTransformerFactory().newTransformer();
-		t.setOutputProperty(OutputKeys.OMIT_XML_DECLARATION, "yes");
+		if (omitXmlDeclaration) {
+			t.setOutputProperty(OutputKeys.OMIT_XML_DECLARATION, "yes");
+		}
 		StringWriter sw = new StringWriter();
 		t.transform(new DOMSource(node), new StreamResult(sw));
 		return sw.toString();