+++ RELEASES.md
@@ -12,7 +12,9 @@ Upcoming
 [Commits](https://github.com/ibissource/iaf/compare/v7.0-B2...HEAD)
 [![Build Status](https://travis-ci.org/ibissource/iaf.png)](https://travis-ci.org/ibissource/iaf)
 
-- Add json to xml and xml to json converstion as well as json validation to xmlvalidators
+- Add json to xml and xml to json conversion as well as json validation to xmlvalidators
+- Prevent XML Entity Expansion (XEE) injection
+- Upgrade ibis-servicedispatcher to version 1.4